pyproject.toml
[project]
name = "first-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.7",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.7",
    "langgraph-api>=0.7.9",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "langgraph-cli>=0.4.12",
    "pytest>=9.0.2",
]
